Move getTokenLink to the bottom of deposit button. Make deposit button width larger. Reorder files
