+++ src/components/earn/PoolCardTri/GetTokenLink/GetTokenLink.constants.ts
